+++ lightsaber/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-schwarz.it.lightsaber.LightsaberSymbolProcessorProvider
+schwarz.it.lightsaber.LightsaberKspProcessorProvider
