+++ lightsaber/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-schwarz.it.lightsaber.LightsaberProcessor
+schwarz.it.lightsaber.LightsaberJavacProcessor
